Datasets/evalplus/Java/src/test/java/com/example/HumanEval_87Test.java
package com.example;

import org.junit.Test;
import static org.junit.Assert.*;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Collections;

public class HumanEval_87Test {

    @Test
    public void test_0() {
        HumanEval_87 instance = new HumanEval_87();
        ArrayList<ArrayList<Integer>> input = new ArrayList<>(Arrays.asList(
                new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 6)),
                new ArrayList<>(Arrays.asList(1, 2, 3, 4, 1, 6)),
                new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 1))
        ));
        ArrayList<Integer[]> expectedResult = new ArrayList<>(Arrays.asList(
                new Integer[]{0, 0},
                new Integer[]{1, 4},
                new Integer[]{1, 0},
                new Integer[]{2, 5},
                new Integer[]{2, 0}
        ));
        assertEquals(expectedResult, instance.getRow(input, 1));
    }

    @Test
    public void test_1() {
        HumanEval_87 instance = new HumanEval_87();
        ArrayList<ArrayList<Integer>> input = new ArrayList<>(Collections.nCopies(6, new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 6))));
        ArrayList<Integer[]> expectedResult = new ArrayList<>(Collections.nCopies(6, new Integer[]{0, 1}));
        assertEquals(expectedResult, instance.getRow(input, 2));
    }

    // Add more tests here according to the list of Python tests...

    @Test
    public void test_3() {
        HumanEval_87 instance = new HumanEval_87();
        ArrayList<ArrayList<Integer>> input = new ArrayList<>();
        ArrayList<Integer[]> expectedResult = new ArrayList<>();
        assertEquals(expectedResult, instance.getRow(input, 1));
    }

    // ...

    @Test
    public void test_12() {
        HumanEval_87 instance = new HumanEval_87();
        ArrayList<ArrayList<Integer>> input = new ArrayList<>(Collections.singletonList(new ArrayList<>(Collections.singletonList(0))));
        ArrayList<Integer[]> expectedResult = new ArrayList<>();
        assertEquals(expectedResult, instance.getRow(input, 1));
    }
}
// End of Test

